Reorganize/refactor file structure
