+++ .eslintcache
@@ -1,1 +1,1 @@
-[{"C:\\Users\\emman\\OneDrive\\Desktop\\WebDev\\TSE\\distribution-web-app\\src\\index.js":"1","C:\\Users\\emman\\OneDrive\\Desktop\\WebDev\\TSE\\distribution-web-app\\src\\App.js":"2","C:\\Users\\emman\\OneDrive\\Desktop\\WebDev\\TSE\\distribution-web-app\\src\\reportWebVitals.js":"3","C:\\Users\\emman\\OneDrive\\Desktop\\WebDev\\TSE\\distribution-web-app\\src\\pages\\AgencyProfileForm.js":"4","C:\\Users\\emman\\OneDrive\\Desktop\\WebDev\\TSE\\distribution-web-app\\src\\pages\\FormComponents\\FormButton.jsx":"5","C:\\Users\\emman\\OneDrive\\Desktop\\WebDev\\TSE\\distribution-web-app\\src\\pages\\FormComponents\\InputText.jsx":"6"},{"size":517,"mtime":1607384692507,"results":"7","hashOfConfig":"8"},{"size":192,"mtime":1607388718637,"results":"9","hashOfConfig":"8"},{"size":375,"mtime":1607384692509,"results":"10","hashOfConfig":"8"},{"size":4542,"mtime":1607388854273,"results":"11","hashOfConfig":"8"},{"size":1033,"mtime":1607388631867,"results":"12","hashOfConfig":"8"},{"size":998,"mtime":1607388632678,"results":"13","hashOfConfig":"8"},{"filePath":"14","messages":"15","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"16"},"axj2im",{"filePath":"17","messages":"18","errorCount":0,"warningCount":1,"fixableErrorCount":0,"fixableWarningCount":0,"source":"19","usedDeprecatedRules":"16"},{"filePath":"20","messages":"21","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"16"},{"filePath":"22","messages":"23","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"24","messages":"25","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"26","messages":"27","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"C:\\Users\\emman\\OneDrive\\Desktop\\WebDev\\TSE\\distribution-web-app\\src\\index.js",[],["28","29"],"C:\\Users\\emman\\OneDrive\\Desktop\\WebDev\\TSE\\distribution-web-app\\src\\App.js",["30"],"import logo from \"./logo.svg\";\r\nimport \"./App.css\";\r\nimport AgencyProfileForm from \"./pages/AgencyProfileForm\";\r\n\r\nfunction App() {\r\n  return <AgencyProfileForm />;\r\n}\r\n\r\nexport default App;\r\n","C:\\Users\\emman\\OneDrive\\Desktop\\WebDev\\TSE\\distribution-web-app\\src\\reportWebVitals.js",[],"C:\\Users\\emman\\OneDrive\\Desktop\\WebDev\\TSE\\distribution-web-app\\src\\pages\\AgencyProfileForm.js",[],"C:\\Users\\emman\\OneDrive\\Desktop\\WebDev\\TSE\\distribution-web-app\\src\\pages\\FormComponents\\FormButton.jsx",[],"C:\\Users\\emman\\OneDrive\\Desktop\\WebDev\\TSE\\distribution-web-app\\src\\pages\\FormComponents\\InputText.jsx",[],{"ruleId":"31","replacedBy":"32"},{"ruleId":"33","replacedBy":"34"},{"ruleId":"35","severity":1,"message":"36","line":1,"column":8,"nodeType":"37","messageId":"38","endLine":1,"endColumn":12},"no-native-reassign",["39"],"no-negated-in-lhs",["40"],"no-unused-vars","'logo' is defined but never used.","Identifier","unusedVar","no-global-assign","no-unsafe-negation"]
+[{"/Users/billwu/Desktop/TSE/distribution-web-app/src/index.js":"1","/Users/billwu/Desktop/TSE/distribution-web-app/src/reportWebVitals.js":"2","/Users/billwu/Desktop/TSE/distribution-web-app/src/App.js":"3","/Users/billwu/Desktop/TSE/distribution-web-app/src/pages/agency-profile-form/AgencyProfileForm.js":"4","/Users/billwu/Desktop/TSE/distribution-web-app/src/pages/agency-profile-form/FormSectionHeader.jsx":"5","/Users/billwu/Desktop/TSE/distribution-web-app/src/pages/agency-profile-form/FormButton.jsx":"6","/Users/billwu/Desktop/TSE/distribution-web-app/src/pages/agency-profile-form/InputDropdown.jsx":"7","/Users/billwu/Desktop/TSE/distribution-web-app/src/pages/agency-profile-form/InputText.jsx":"8"},{"size":500,"mtime":1607383180057,"results":"9","hashOfConfig":"10"},{"size":362,"mtime":1607383180058,"results":"11","hashOfConfig":"10"},{"size":203,"mtime":1608003539897,"results":"12","hashOfConfig":"10"},{"size":4465,"mtime":1608003849463,"results":"13","hashOfConfig":"10"},{"size":769,"mtime":1608003594832,"results":"14","hashOfConfig":"10"},{"size":988,"mtime":1608003663756,"results":"15","hashOfConfig":"10"},{"size":1110,"mtime":1608003685539,"results":"16","hashOfConfig":"10"},{"size":959,"mtime":1608003625831,"results":"17","hashOfConfig":"10"},{"filePath":"18","messages":"19","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"20"},"180g3n5",{"filePath":"21","messages":"22","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"20"},{"filePath":"23","messages":"24","errorCount":0,"warningCount":1,"fixableErrorCount":0,"fixableWarningCount":0,"source":"25","usedDeprecatedRules":"20"},{"filePath":"26","messages":"27","errorCount":0,"warningCount":1,"fixableErrorCount":0,"fixableWarningCount":0,"source":null},{"filePath":"28","messages":"29","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"20"},{"filePath":"30","messages":"31","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"20"},{"filePath":"32","messages":"33","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"20"},{"filePath":"34","messages":"35","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"20"},"/Users/billwu/Desktop/TSE/distribution-web-app/src/index.js",[],["36","37"],"/Users/billwu/Desktop/TSE/distribution-web-app/src/reportWebVitals.js",[],"/Users/billwu/Desktop/TSE/distribution-web-app/src/App.js",["38"],"import logo from \"./logo.svg\";\nimport \"./App.css\";\nimport AgencyProfileForm from \"./pages/agency-profile-form/AgencyProfileForm\";\n\nfunction App() {\n  return <AgencyProfileForm />;\n}\n\nexport default App;\n","/Users/billwu/Desktop/TSE/distribution-web-app/src/pages/agency-profile-form/AgencyProfileForm.js",["39"],"/Users/billwu/Desktop/TSE/distribution-web-app/src/pages/agency-profile-form/FormSectionHeader.jsx",[],"/Users/billwu/Desktop/TSE/distribution-web-app/src/pages/agency-profile-form/FormButton.jsx",[],"/Users/billwu/Desktop/TSE/distribution-web-app/src/pages/agency-profile-form/InputDropdown.jsx",[],"/Users/billwu/Desktop/TSE/distribution-web-app/src/pages/agency-profile-form/InputText.jsx",[],{"ruleId":"40","replacedBy":"41"},{"ruleId":"42","replacedBy":"43"},{"ruleId":"44","severity":1,"message":"45","line":1,"column":8,"nodeType":"46","messageId":"47","endLine":1,"endColumn":12},{"ruleId":"44","severity":1,"message":"48","line":7,"column":8,"nodeType":"46","messageId":"47","endLine":7,"endColumn":21},"no-native-reassign",["49"],"no-negated-in-lhs",["50"],"no-unused-vars","'logo' is defined but never used.","Identifier","unusedVar","'InputDropdown' is defined but never used.","no-global-assign","no-unsafe-negation"]
