+++ src/App.js
@@ -1,6 +1,6 @@
 import logo from "./logo.svg";
 import "./App.css";
-import AgencyProfileForm from "./pages/AgencyProfileForm";
+import AgencyProfileForm from "./pages/agency-profile-form/AgencyProfileForm";
 
 function App() {
   return <AgencyProfileForm />;
