+++ src/pages/agency-profile-form/formstyle.css
@@ -1,5 +1,3 @@
-
-
 .page-header {
   position: absolute;
